Fix for Bug #3: Rubble is Not Bulldozable
Fixed an issue where rubble tiles could not be cleared using the bulldozer tool.
The problem was caused by rubble tiles being marked as `"noburn"` instead of `"bulldozable"` in `tiles.rc` (lines 59-62).
Updated the tile definitions to allow rubble to be bulldozed properly.
Tested in-game and confirmed that rubble can now be cleared as expected.